requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework~=3.12.4
django-cors-headers~=3.10.0
simplejson~=3.17.3
requests~=2.25.1